physical_ai_k3s/apps/central-hub/requirements.txt
flask
kubernetes
pyyaml
redis
flask
kubernetes
pyyaml
redis
google-generativeai